requirements.txt
Flask
psycopg2-binary
bcrypt
PyJWT
flask-mail
python-dotenv
pytz